Added behaviour blending by weight.
