+++ Assets/Prefabs/Monsters/SimpleMonster.prefab
@@ -335,6 +335,7 @@ GameObject:
   - component: {fileID: -8664493499008248624}
   - component: {fileID: 4170874774547565885}
   - component: {fileID: 1262647091002150027}
+  - component: {fileID: 8877848422801390066}
   m_Layer: 0
   m_Name: SimpleMonster
   m_TagString: Enemy
@@ -550,6 +551,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c983f7a3bdf77a04686eb8be607bec57, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  weight: 1
+  target: {fileID: 4305704709854959933, guid: 0b38f5738ea76fc47bbc9479eb38ab82, type: 3}
+--- !u!114 &8877848422801390066
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2873645921850147136}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 66c1d796790c0c0408260dacf14ea289, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  weight: 1
   target: {fileID: 4305704709854959933, guid: 0b38f5738ea76fc47bbc9479eb38ab82, type: 3}
 --- !u!1 &4183647638066128296
 GameObject:
